+++ CHANGES.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and
 this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- EditorQuill supports custom parchment formats with `blockConfig.registerFormats`
+  - see example in [editor-quill/index.stories.js](./src/components/editor-quill/index.stories.js)
+
+### Changed
+
+- EditorQuill does not render custom toolbar, when `toolbarOptions` is provided
+  in the `blockConfig`
+
+### Removed
+
+- Removed `hideToolbarOnBlur` feature of EditorQuill, because of [react-quill/issues/276](https://github.com/zenoamaro/react-quill/issues/276#issuecomment-384618178)
+
 ## 2018/12/10 [1.2.0][12]
 
 ### Added